Moved code from python to asr folder
Also added the missing example WAV file.
