+++ asr/dataset/common_voice_loader.py
@@ -14,12 +14,12 @@ from multiprocessing import Pool, Lock, cpu_count
 from scipy.io import wavfile
 from tqdm import tqdm
 
-from python.dataset import download
-from python.dataset.config import CACHE_DIR, CORPUS_DIR, sox_commandline
-from python.dataset.config import CSV_HEADER_PATH, CSV_HEADER_LABEL, CSV_HEADER_LENGTH
-from python.dataset.csv_file_helper import generate_csv
-from python.params import MIN_EXAMPLE_LENGTH, MAX_EXAMPLE_LENGTH
-from python.util.storage import delete_file_if_exists
+from asr.dataset import download
+from asr.dataset.config import CACHE_DIR, CORPUS_DIR, sox_commandline
+from asr.dataset.config import CSV_HEADER_PATH, CSV_HEADER_LABEL, CSV_HEADER_LENGTH
+from asr.dataset.csv_file_helper import generate_csv
+from asr.params import MIN_EXAMPLE_LENGTH, MAX_EXAMPLE_LENGTH
+from asr.util.storage import delete_file_if_exists
 
 # Path to the Mozilla Common Voice dataset.
 __URL = 'https://common-voice-data-download.s3.amazonaws.com/cv_corpus_v1.tar.gz'
