+++ asr/dataset/csv_file_helper.py
@@ -6,11 +6,11 @@ import csv
 import os
 import re
 
-from python.dataset.config import CSV_HEADER_LABEL, CSV_HEADER_LENGTH, CSV_FIELDNAMES
-from python.dataset.config import LABEL_WHITELIST_PATTERN, CSV_DIR, CSV_DELIMITER
-from python.input_functions import WIN_STEP
-from python.util import storage
-from python.util.matplotlib_helper import pyplot_display
+from asr.dataset.config import CSV_HEADER_LABEL, CSV_HEADER_LENGTH, CSV_FIELDNAMES
+from asr.dataset.config import LABEL_WHITELIST_PATTERN, CSV_DIR, CSV_DELIMITER
+from asr.input_functions import WIN_STEP
+from asr.util import storage
+from asr.util.matplotlib_helper import pyplot_display
 
 
 def generate_csv(dataset_name, target, csv_data):
