+++ asr/dataset/sd_estimator.py
@@ -10,8 +10,8 @@ from multiprocessing import Pool, Lock, cpu_count
 import numpy as np
 from tqdm import tqdm
 
-from python.input_functions import load_sample
-from python.params import BASE_PATH
+from asr.input_functions import load_sample
+from asr.params import BASE_PATH
 
 __DATASETS_PATH = os.path.join(BASE_PATH, '../datasets/speech_data')
 __FEATURE_TYPE = 'mel'
