+++ asr/dataset/timit_loader.py
@@ -9,10 +9,10 @@ import os
 
 from scipy.io import wavfile
 
-from python.dataset.config import CORPUS_DIR
-from python.dataset.config import CSV_HEADER_PATH, CSV_HEADER_LABEL, CSV_HEADER_LENGTH
-from python.dataset.csv_file_helper import generate_csv
-from python.params import MIN_EXAMPLE_LENGTH, MAX_EXAMPLE_LENGTH
+from asr.dataset.config import CORPUS_DIR
+from asr.dataset.config import CSV_HEADER_PATH, CSV_HEADER_LABEL, CSV_HEADER_LENGTH
+from asr.dataset.csv_file_helper import generate_csv
+from asr.params import MIN_EXAMPLE_LENGTH, MAX_EXAMPLE_LENGTH
 
 # Path to the TIMIT dataset.
 __NAME = 'timit'
