+++ asr/util/tf_contrib.py
@@ -4,7 +4,7 @@ Utility and helper methods for TensorFlow speech learning.
 
 import tensorflow as tf
 
-from python.params import FLAGS, TF_FLOAT
+from asr.params import FLAGS, TF_FLOAT
 
 
 class AdamOptimizerLogger(tf.train.AdamOptimizer):
